+++ chaos/src/controller/ControllerRaw.hpp
@@ -42,7 +42,7 @@ namespace Chaos {
     std::shared_ptr<ControllerState> mControllerState;
     int mLastFactoryVendor = -1;
     int mLastFactoryProduct = -1;
-    std::uint64_t mLastTransportGeneration = 0;
+    std::uint32_t mLastTransportGeneration = 0;
 
     void initializeControllerStateIfPossible();
