+++ chaos/src/controller/usb_sniffify/include/raw-gadget.hpp
@@ -69,14 +69,14 @@ public:
   bool readyProductVendor();
   int getVendor();
   int getProduct();
-  std::uint64_t getConnectionGeneration() const;
+  std::uint32_t getConnectionGeneration() const;
   void requestReconnect();
   
 private:
   int product = 0;
   int vendor = 0;
   std::atomic<bool> haveProductVendor{false};
-  std::atomic<std::uint64_t> connectionGeneration{0};
+  std::atomic<std::uint32_t> connectionGeneration{0};
   std::vector<EndpointObserver*> observers;
   
   std::atomic<bool> keepRunning{false};
